+++ src/main/java/persistence/sql/dml/query/metadata/WhereCondition.java
@@ -1,4 +1,4 @@
-package persistence.sql.dml;
+package persistence.sql.dml.query.metadata;
 
 public record WhereCondition(ColumnName columnName,
                             String operator) {
