+++ src/main/java/persistence/validator/AnnotationValidator.java
@@ -1,4 +1,4 @@
-package persistence.sql.dml;
+package persistence.validator;
 
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Field;
